Fixed workspace-copilot PR + cleaned up some code and project settings
