+++ .pylintrc
@@ -4,7 +4,7 @@ max-line-length=120
 min-public-methods=0
 
 [MESSAGES CONTROL]
-# Disable the message "C0114: Missing module docstring"
+# Disabled the message "C0114: Missing module docstring"
 # Disabled the message "W0511: fixme"
 # Disabled the message "R0913: Too many arguments"
 # Disabled the message "W0212: Access to a protected member _ of a client class"
