+++ bot/botcommands/__init__.py
@@ -3,3 +3,4 @@
 from .bot_configs import BotConfigsCog
 from .member import MemberCog
 from .moderation import ModerationCog
+from .reaction_role import ReactionRoleCog
