+++ bot/db/models/__init__.py
@@ -2,4 +2,5 @@
 
 from .bot_configs import GlobalConfig, SpamConfigs
 from .entity import Entity
+from .reaction_role import ReactionRole
 from .user import AdeptMember
