+++ bot/db/services/__init__.py
@@ -2,4 +2,5 @@
 
 from .base_service import BaseService
 from .configs_service import ConfigsService
+from .reaction_role_service import ReactionRoleService
 from .user_service import UserService
