+++ pyproject.toml
@@ -1,0 +1,6 @@
+[tool.isort]
+profile = "black"
+
+[tool.black]
+line-length = 120
+target-version = ["py313"]
