+++ requirements.txt
@@ -1,8 +1,4 @@
 discord.py[speed]<3.0.0
 jsonpickle
 pymongo[srv]<5.0.0
-
-# Linter
-pylint
-# Formatter
-black
+audioop-lts
